feat(core): Implement inheritance for ProductVariant.trackInventory
The value can now be set to inherit from the GlobalSettings.trackInventory value.

BREAKING CHANGE: The `ProductVariant.trackInventory` field is now an Enum rather than a boolean, allowing explicit inheritance of the value set in GlobalSettings. This will require a DB migration with a custom query to transform the previous boolean values to the new enum (string) values of "TRUE", "FALSE" or "INHERIT". Check the release blog post for more details.