feat(core): Re-work handling of taxes, order-level discounts
Relates to #573. This commit aims to correct the way that taxes are handled for Order-level
promotions. This is done by prorating (proportionally-distributing) Order-level discounts among
the OrderItems of the Order. Implementing this involved several DB schema changes as well as
a general overhaul of the GraphQL API as relates to order pricing.

BREAKING CHANGE: There have been some major changes to the way that Order taxes and discounts are
handled. For a full discussion of the issues behind these changes see #573. These changes will
require a DB migration as well as possible custom scripts to port existing Orders to the new
format. See the release blog post for details.

The following GraphQL `Order` type properties have changed:

* `subTotalBeforeTax` has been removed, `subTotal` now excludes tax, and
`subTotalWithTax` has been added.
* `totalBeforeTax` has been removed, `total` now excludes tax, and
`totalWithTax` has been added.