fix(elastic-search-plugin): Refactor indexing to allow searching by fields on all variants (#1086)
Closes #1018, closes #1017 

This commit changes the way products and variants are indexed in Elasticsearch. Previously we used 2 separate indices - one for Products and one for ProductVariants. This caused issues like #1017 which could not be overcome without changing the way we index. Now we have a single index and use the Elasticsearch "collapse" feature in order to do the grouping of variants into products.

This simplifies the logic for indexing quite a bit, and also speeds up indexing. The downside is that querying is slightly less efficient now, since a separate query is now required in order to get the total results count. However, this slight performance hit is worth it for having correct functioning, which was not possible before.