feat(core): Use variant featuredAsset in OrderLine if available
Relates to #1488. This might technically be viewed as a breaking change, but in another sense it is
finally implementing what seems to be the expected behaviour per an
informal poll I conducted on Slack where 7 of 7 responded that it should
work this way. To me it seems that this is OK to put in a minor release and should not wait until
v2.