fix(core): Fix regression when querying custom field relations
Fixes #1664. Relates to #1636 because it partially reverts that fix. This is
unfortunate but having the GraphQL API work without throwing errors is more
important than not loading custom fields in the service layer.

To mitigate the effect of the partial reversion of the #1636 fix, I've added
a debug-level log message to alert to the fact that an expected relation
could not be loaded.